feat: add webtask that queries twitter for dad jokes and sends them over sms
I developed this in the auth0 webtask web editor, so this repo is still missing some things like the package.json if you were going to initialize using the webtask cli. You'll need to define some secrets like the twitter and twilio creds, src and dest phone numbers, as well as a basic hard-coded api token for a single user auth.